chore: Improve DWO Webhook
